Make ExtendedPath ako FilePath
